refactor(textarea): separation of autosize processing into useAutosize
